LPD-8519 Remove service from KBArchiveFactory, rename it to KBArchiveFactoryUtil and pass ConfigurationProvider as parameter for the createKBArchive method
