+++ modules/apps/knowledge-base/knowledge-base-service/src/main/java/com/liferay/knowledge/base/internal/importer/util/KBArchiveFactoryUtil.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.knowledge.base.internal.importer;
+package com.liferay.knowledge.base.internal.importer.util;
 
 import com.liferay.knowledge.base.configuration.KBGroupServiceConfiguration;
 import com.liferay.knowledge.base.constants.KBConstants;
 import com.liferay.knowledge.base.exception.KBArticleImportException;
+import com.liferay.knowledge.base.internal.importer.KBArchive;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.module.configuration.ConfigurationProvider;
@@ -29,16 +30,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
 /**
  * @author Adolfo Pérez
  */
-@Component(service = KBArchiveFactory.class)
-public class KBArchiveFactory {
+public class KBArchiveFactoryUtil {
 
-	public KBArchive createKBArchive(long groupId, ZipReader zipReader)
+	public static KBArchive createKBArchive(
+			ConfigurationProvider configurationProvider, long groupId,
+			ZipReader zipReader)
 		throws PortalException {
 
 		List<String> entries = zipReader.getEntries();
@@ -49,7 +48,7 @@ public class KBArchiveFactory {
 		}
 
 		KBGroupServiceConfiguration kbGroupServiceConfiguration =
-			_configurationProvider.getConfiguration(
+			configurationProvider.getConfiguration(
 				KBGroupServiceConfiguration.class,
 				new GroupServiceSettingsLocator(
 					groupId, KBConstants.SERVICE_NAME));
@@ -87,9 +86,6 @@ public class KBArchiveFactory {
 
 		return new KBArchiveImpl(kbArchiveState.getFolders());
 	}
-
-	@Reference
-	private ConfigurationProvider _configurationProvider;
 
 	private static final class FileImpl implements KBArchive.File {
